requirements.txt
Flask==3.1.0
requests==2.31.0
beautifulsoup4==4.13.3
python-dotenv==0.21.1
whois==0.9.26